Move the windows login window and control to the RoadCaptain.App.Windows project
