+++ src/RoadCaptain.App.Runner/DelegateDecorator.cs
@@ -5,6 +5,7 @@ using Avalonia.Controls.ApplicationLifetimes;
 using Avalonia.Threading;
 using RoadCaptain.App.Runner.Models;
 using RoadCaptain.App.Runner.ViewModels;
+using RoadCaptain.App.Shared.Models;
 
 namespace RoadCaptain.App.Runner
 {
