+++ src/RoadCaptain.App.Runner/RoadCaptain.App.Runner.csproj
@@ -9,7 +9,6 @@
 		<RuntimeIdentifier Condition="$([MSBuild]::IsOSPlatform('WINDOWS'))">win-x64</RuntimeIdentifier>
 		<!-- Need to specify both RuntimeIdentifier and RuntimeIdentifiers because of MSBuild weirdness -->
 		<RuntimeIdentifiers Condition="$([MSBuild]::IsOSPlatform('WINDOWS'))">win-x64</RuntimeIdentifiers>
-		<UseWindowsForms Condition="$([MSBuild]::IsOSPlatform('WINDOWS'))">true</UseWindowsForms>
 
 		<OutputType>WinExe</OutputType>
 		<OutputType Condition="'$(RuntimeIdentifier)'=='linux-x64' Or '$(RuntimeIdentifier)'=='osx-x64'">Exe</OutputType>
@@ -49,7 +48,6 @@
 		<!--Condition below is needed to remove Avalonia.Diagnostics package from build output in Release configuration.-->
 		<PackageReference Condition="'$(Configuration)' == 'Debug'" Include="Avalonia.Diagnostics" Version="0.10.12" />
 		<PackageReference Include="Avalonia.ReactiveUI" Version="0.10.12" />
-		<PackageReference Include="Microsoft.Web.WebView2" Version="1.0.1210.39" Condition="$([MSBuild]::IsOSPlatform('WINDOWS'))" /> <!--Only on Windows -->
 		<PackageReference Include="WebViewControl-Avalonia" Version="2.91.23" Condition="!$([MSBuild]::IsOSPlatform('WINDOWS'))" /> <!--_not_ on Windows-->
 		<PackageReference Include="XamlNameReferenceGenerator" Version="1.3.4" />
 		<PackageReference Include="Autofac" Version="6.3.0" />
