+++ src/RoadCaptain.App.Shared/Views/ZwiftLoginWindowBase.cs
@@ -1,0 +1,10 @@
+﻿using Avalonia.Controls;
+using RoadCaptain.App.Shared.Models;
+
+namespace RoadCaptain.App.Shared.Views
+{
+    public abstract class ZwiftLoginWindowBase : Window
+    {
+        public TokenResponse? TokenResponse { get; protected set; }
+    }
+}
