+++ src/RoadCaptain.App.Windows/Controls/WinWebView.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.IO;
-using Avalonia.Controls;
+﻿using Avalonia.Controls;
 using Avalonia.Platform;
 using Microsoft.Web.WebView2.Core;
 using Microsoft.Web.WebView2.WinForms;
