+++ src/RoadCaptain.App.Windows/WindowsModule.cs
@@ -1,6 +1,8 @@
 ﻿using Autofac;
 using RoadCaptain.App.Shared.UserPreferences;
+using RoadCaptain.App.Shared.Views;
 using RoadCaptain.App.Windows.UserPreferences;
+using RoadCaptain.App.Windows.Views;
 
 namespace RoadCaptain.App.Windows
 {
@@ -9,6 +11,7 @@ namespace RoadCaptain.App.Windows
         protected override void Load(ContainerBuilder builder)
         {
             builder.RegisterType<WindowsUserPreferences>().As<IUserPreferences>().SingleInstance();
+            builder.RegisterType<ZwiftLoginWindowPlatformWin>().As<ZwiftLoginWindowBase>();
         }
     }
 }
